ectool: Adapt to match mainline ioctl API
The mainline kernel has a different ioctl API where
fixed-size arrays are used instead of pointers for
the in and out buffers to avoid the need of having
a compat ioctl interface for 32-bit user-space in
a 64-bit kernel.

Modify the data structures used as arguments in
ioctl commands accordingly.